requirements.txt
crewai
openai
requests
python-dotenv
